edit readme

+++ pkg/__main__.py
@@ -2,8 +2,10 @@ import os
 import sys
 import time
 
-from router import *
-from fetcher import *
+from pkg.router import *
+from pkg.fetcher import *
+
+
 
 headers = {
     'User-Agent': 'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_12_3) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/56.0.2924.87 Safari/537.36',
@@ -37,22 +39,32 @@ def write_file(title, content, folder_name):
         else:
             name += char
     fname = './' + folder_name + '/' + name + '.txt'
+    if os.path.isfile(fname):
+        print(fname + ' already exists')
+        return
     with open(fname, 'w', encoding= 'utf-8') as f:
         f.write(content)
 
 def dl_conbined(articles):
     """
     type articles: List[String] or None
+    rtype: int, number of articles downloaded
     """
+    cnt = 0
     if articles:
         for article in articles:
             returned_item = get_content(article, zh_cn_only= zhcn_only)
             if returned_item:
                 content, title = returned_item
                 write_file(title, content, folder_name)
-                time.sleep(2)
+                # success count + 1
+                cnt += 1
+    return cnt     
 
 class Logger(object):
+    """
+    Logging module
+    """
     def __init__(self, filename, stream=sys.stdout):
 	    self.terminal = stream
 	    self.log = open(filename, "wb", buffering=0)
@@ -60,7 +72,8 @@ class Logger(object):
     def write(self, *message):
         message = ",".join([str(it) for it in message])
         self.terminal.write(str(message))
-        self.log.write(str(message).encode('utf-8'))
+        prefix = '[' + time.strftime("%Y-%m-%d %H:%M:%S", time.localtime()) + ']'
+        self.log.write(prefix.encode('utf-8') + str(message).encode('utf-8'))
 
     def flush(self):
         pass
@@ -69,30 +82,38 @@ def time_helper(seperator = '_', to_sec = True):
     """
     return a string like 2020_09_11_22_43_00 (if to_sec is True) or 2020_09_11_22_43 (if to_sec is False)
     """
-    localtime = time.asctime(time.localtime(time.time()))
     if to_sec:
         return time.strftime("%Y" + seperator + "%m" + seperator + "%d" + seperator + "%H" + seperator + "%M" + seperator + "%S", time.localtime()) 
     return time.strftime("%Y" + seperator + "%m" + seperator + "%d" + seperator + "%H" + seperator + "%M", time.localtime()) 
 
 if __name__ == '__main__':
-    start_time = time_helper('-')
+    # how many articles downloaded
+    article_cnt = 0
+    # logging
     if not os.path.exists('./log'):
         os.mkdir('./log')
+    start_time = time_helper()
     sys.stdout = Logger('./log/' + start_time + '.log')
-    print("Start time:" + start_time)
+
     print("Cannot be used in mainland China, please use a VPN with global mode or ask someone overseas to help you")
     print("All comments are written in English, some machines cannot print simplified Chinese on console")
-
-    url = input('Please paste an AO3 url here:  ')
-    zhcn_only = input("Simplified Chinese only?(Y if yes else press anything else): ")
+    print("Please paste an AO3 url here:  ")
+    url = input()
+    print("Your option is " + url)
+    print("Simplified Chinese only?(Y if yes else press anything else): ")
+    zhcn_only = input()
+    print("Your option is " + zhcn_only)
     if zhcn_only == 'y' or zhcn_only == 'Y':
         zhcn_only = True
     else:
         zhcn_only = False
+    # check if the input folder name is valid in Windows
     valid = False
     while valid == False:
         valid = True
-        folder_name = input("Input a folder name to store downloaded file: ")
+        print("Input a folder name to store downloaded file: ")
+        folder_name = input()
+        print("Your option is " + folder_name)
         win_invalid_chars = "\\/:*?\"<>|"
         for char in folder_name:
             if char in win_invalid_chars:
@@ -105,22 +126,31 @@ if __name__ == '__main__':
     if not os.path.exists('./' + folder_name):
         os.mkdir('./' + folder_name)
     url = url.rstrip('/')
-    if not url.startswith("https://archiveofourown.org/"):
-        print("Please use url starts with \'https://archiveofourown.org/\'")
+    url = url.lstrip()
+    option_1 = url.startswith("https://archiveofourown.org") 
+    option_2 = url.startswith("https://www.archiveofourown.org")
+    if option_1 == False and option_2 == False:
+        print("Please use url starts with \'https://archiveofourown.org/\' or \'https://www.archiveofourown.org/\'")
         exit(0)
-    print("Process started, connecting to the website...")
+    print("Process started, connecting to the website...if connection failed or it is too slow, please check if you have pyopenssl in your environment")
+    
     try:
         req = requests.get(url, headers = headers)
         html = req.text
+        print("connection established")
     except:
         print("Cannot visit the url you pasted, please check your internet or VPN settings")
-        exit(1)
-
+        exit(0)
+    if not req:
+        print("request of url " + url + " not found")
+        exit(0)
+    # single article
     if is_article(url):
         returned_item = get_content(url, zh_cn_only= zhcn_only)
         if returned_item:
             content, title = returned_item
             write_file(title, content, folder_name)
+            article_cnt += 1
 
     else:
         soup = BeautifulSoup(html, 'html.parser')
@@ -130,18 +160,25 @@ if __name__ == '__main__':
             urls, curpage = page_detail
             page_amount = len(urls)
             print(str(page_amount) + " related pages detected, current page index is " + curpage + ", do you want to download all pages? Page index from 1 to " + str(page_amount))
-            argument = input("press key a to download all, press s to select start and and index, press other keys to download articles in the current page ")
+            print("press key a to download all, press s to select start and and index, press other keys to download articles in the current page ")
+            argument = input()
+            print("Your option is " + argument)
+            # download all related pages
             if argument == 'a' or argument == 'A':
                 for url in urls:
                     articles = get_articles(url)
-                    dl_conbined(articles)
-                    time.sleep(60)
+                    article_cnt += dl_conbined(articles)
+            # download page from a range
             elif argument == 's' or argument == 'S':
                 start = None
                 end = None
                 while not start or not end:
-                    start = input("Start: ")
-                    end = input("End: ")
+                    print("Start: ")
+                    start = input()
+                    print("The start page you choose is " + start)
+                    print("End: ")
+                    end = input()
+                    print("The end page you choose is " + end)
                     if start.isdecimal() == False or end.isdecimal() == False:
                         print("Please input number for start and end index, input again")
                         start, end = None, None
@@ -155,15 +192,18 @@ if __name__ == '__main__':
                 for i in range(start, end):
                     url = urls[i]
                     articles = get_articles(url)
-                    dl_conbined(articles)
-                    time.sleep(60)
+                    article_cnt += dl_conbined(articles)
+            # current page only
             else:
                 url = urls[int(curpage) - 1]
                 articles = get_articles(url)
-                dl_conbined(articles)
-    end_time = time_helper('-')
-    print("End time:" + end_time)
+                article_cnt += dl_conbined(articles)
+        else:
+            articles = get_articles(url)
+            article_cnt += dl_conbined(articles)
+    print(str(article_cnt) + ' articles downloaded')
     print("Download Finished, if you cannot find the text file in that folder, please close simplified Chinese only option")
-    input("Press any key to close the program")
+    print("Press any key to close the program")
+    input()
